pyproject.toml
[project]
name = "flatlander"
version = "0.1.0"
description = "Fit complex images with simple geometric shapes."
authors = [{ name = "weilycoder", email = "weily2009@outlook.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=12.1.1,<13.0.0)",
    "rich (>=14.3.3,<15.0.0)",
    "svgwrite (>=1.4.3,<2.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
